refactor: extract pure functions from utils module
